requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
pip==23.2.1
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
